+++ Release Notes.md
@@ -8,6 +8,7 @@
 * `FlipView` has a new `isFlipped` binding, so that it can be read from outside as well.
 * `FormText` now supports iOS 13.
 * `FormText` has a new `trailingView` property. 
+* There is a new `UIImage+Rotated` extension.